+++ R/as-tsibble.R
@@ -5,7 +5,7 @@ globalVariables(c("key", "value", "zzz"))
 #' @param ... A set of name-value pairs. The names of "key" and "index" should
 #' be avoided as they are used as the arguments.
 #' @param key Structural variable(s) that define unique time indices, used with
-#' the helper [id]. If a univariate time series (without an explicit key), 
+#' the helper [id]. If a univariate time series (without an explicit key),
 #' simply call `id()`. See below for details.
 #' @param index A bare (or unquoted) variable to specify the time index variable.
 #' @param regular Regular time interval (`TRUE`) or irregular (`FALSE`). `TRUE`
@@ -17,7 +17,7 @@ globalVariables(c("key", "value", "zzz"))
 #'
 #' @inheritSection tsibble-package Interval
 #'
-#' @details A valid tsibble does not arrange the index in the time order. 
+#' @details A valid tsibble does not arrange the index in the time order.
 #' Please use [arrange] to get the order by time.
 #'
 #' @return A tsibble object.
@@ -29,27 +29,29 @@ globalVariables(c("key", "value", "zzz"))
 #'   value = rnorm(10),
 #'   key = id(), index = date
 #' )
+#' tbl1
 #'
 #' # create a tsibble with one key ----
 #' tbl2 <- tsibble(
-#'   qtr = rep(yearquarter(seq(2010, 2012.25, by = 1/ 4)), 3),
+#'   qtr = rep(yearquarter(seq(2010, 2012.25, by = 1 / 4)), 3),
 #'   group = rep(c("x", "y", "z"), each = 10),
 #'   value = rnorm(30),
 #'   key = id(group), index = qtr
 #' )
+#' tbl2
 #'
 #' @export
 tsibble <- function(..., key = id(), index, regular = TRUE) {
   tbl <- tibble::tibble(...)
   index <- enquo(index)
-  tsibble_tbl(tbl, key = key, index = index, regular = regular)
+  as_tsibble(tbl, key = key, index = !! index, regular = regular)
 }
 
 #' Coerce to a tsibble object
 #'
 #' @param x Other objects to be coerced to a tsibble (`tbl_ts`).
 #' @param key Structural variable(s) that define unique time indices, used with
-#' the helper [id]. If a univariate time series (without an explicit key), 
+#' the helper [id]. If a univariate time series (without an explicit key),
 #' simply call `id()`.See below for details.
 #' @param index A bare (or unquoted) variable to specify the time index variable.
 #' @param regular Regular time interval (`TRUE`) or irregular (`FALSE`). `TRUE`
@@ -57,7 +59,7 @@ tsibble <- function(..., key = id(), index, regular = TRUE) {
 #' @param validate `TRUE` suggests to verify that each key or each combination
 #' of key variables lead to unique time indices (i.e. a valid tsibble). It will
 #' also make sure that the nested variables are arranged from lower level to
-#' higher, if nested variables are passed to `key`. If you are sure that it's a 
+#' higher, if nested variables are passed to `key`. If you are sure that it's a
 #' valid input, specify `FALSE` to skip the checks.
 #' @param ... Other arguments passed on to individual methods.
 #'
@@ -105,6 +107,24 @@ as_tsibble.tbl_df <- function(
 
 #' @rdname as-tsibble
 #' @export
+as_tsibble.tbl_ts <- function(
+  x, key = id(), index, regular = TRUE, validate = TRUE, ...
+) {
+  if (is_empty(key)) {
+    key <- key(x)
+  }
+  index <- enquo(index)
+  if (quo_is_missing(index)) {
+    index <- index(x)
+  }
+  tsibble_tbl(
+    x, key = key, index = index, regular = regular,
+    validate = validate
+  )
+}
+
+#' @rdname as-tsibble
+#' @export
 as_tsibble.data.frame <- as_tsibble.tbl_df
 
 #' @keywords internal
@@ -117,35 +137,52 @@ as_tsibble.list <- as_tsibble.tbl_df
 
 #' @keywords internal
 #' @export
-as_tsibble.grouped_ts <- function(
+as_tsibble.grouped_df <- function(
   x, key = id(), index, groups = id(), regular = TRUE, validate = TRUE, ...
 ) {
+  # convert grouped_df to tsibble:
+  # the `groups` arg must be supplied, otherwise returns a `tbl_ts` not grouped
   index <- enquo(index)
+
   tbl <- tsibble_tbl(
     x, key = key, index = index, regular = regular,
     validate = validate
   )
+
   if (is_empty(groups)) {
-    groups <- groups(x)
+    return(tbl)
   }
+
+  groups <- validate_key(x, groups)
+  tbl <- validate_nested(data = tbl, key = groups)
+
   flat_grps <- flatten_key(groups)
-  grped_df <- dplyr::grouped_df(tbl, flat_grps)
+  grped_df <- grouped_df(tbl, flat_grps)
   tibble::new_tibble(
-    tbl, 
+    grped_df,
     "vars" = structure(groups, class = "vars"),
-    "indices" = group_size(grped_df),
     subclass = c("grouped_ts", "tbl_ts")
   )
 }
 
 #' @keywords internal
 #' @export
-as_tsibble.grouped_df <- as_tsibble.grouped_ts
+as_tsibble.grouped_ts <- as_tsibble.grouped_df
 
 #' @keywords internal
 #' @export
 as_tsibble.default <- function(x, ...) {
-  abort("as_tsibble doesn't know how to deal with this type of class yet.")
+  cls <- class(x)[1]
+  msg <- sprintf(
+    "`as_tsibble()` doesn't know how to coerce the `%s` class yet.", cls
+  )
+  abort(msg)
+}
+
+#' @keywords internal
+#' @export
+as_tsibble.NULL <- function(x, ...) {
+  abort("A tsibble must not be NULL.")
 }
 
 #' Return key and measured variables
@@ -210,7 +247,11 @@ key_size <- function(x) {
 
 #' @export
 key_size.tbl_ts <- function(x) {
-  attr(x, "key_indices")
+  key_indices <- attr(x, "key_indices")
+  if (is_empty(key_indices)) {
+    return(NROW(x))
+  }
+  vapply(key_indices, length, integer(1))
 }
 
 #' @rdname key-size
@@ -235,8 +276,8 @@ group_vars.tbl_ts <- function(x) {
 }
 
 #' @export
-group_size.tbl_ts <- function(x) {
-  attr(x, "indices")
+group_size.grouped_ts <- function(x) {
+  vapply(attr(x, "indices"), length, integer(1))
 }
 
 #' @export
@@ -245,8 +286,8 @@ n_groups.tbl_ts <- function(x) {
 }
 
 #' @export
-group_indices.grouped_ts <- function(x) {
-  group_indices(as_tibble(x))
+group_indices.grouped_ts <- function(.data, ...) {
+  attr(.data, "indices")
 }
 
 #' @rdname key
@@ -275,7 +316,7 @@ measured_vars.tbl_ts <- function(x) {
 #' @export
 interval <- function(x) {
   if (is_false(is_tsibble(x))) {
-    abort(paste(expr_label(substitute(x)), "is not a tsibble."))
+    abort(sprintf("%s is not a tsibble.", expr_label(substitute(x))))
   }
   attr(x, "interval")
 }
@@ -284,7 +325,7 @@ interval <- function(x) {
 #' @export
 index <- function(x) {
   if (is_false(is_tsibble(x))) {
-    abort(paste(expr_label(substitute(x)), "is not a tsibble."))
+    abort(sprintf("%s is not a tsibble.", expr_label(substitute(x))))
   }
   attr(x, "index")
 }
@@ -305,7 +346,7 @@ index <- function(x) {
 #' @export
 is_regular <- function(x) {
   if (is_false(is_tsibble(x))) {
-    abort(paste(expr_label(substitute(x)), "is not a tsibble."))
+    abort(sprintf("%s is not a tsibble.", expr_label(substitute(x))))
   }
   attr(x, "regular")
 }
@@ -363,17 +404,27 @@ as.tsibble <- function(x, ...) {
 ## tsibble is a special class of tibble that handles temporal data. It
 ## requires a sequence of time index to be unique across every identifier.
 tsibble_tbl <- function(x, key, index, regular = TRUE, validate = TRUE) {
+  if (NROW(x) == 0 || has_length(x[[1]], 0)) { # no elements or length of 0
+    abort("A tsibble must not be empty.")
+  }
+  # if key is quosures
   use_id(key)
 
-  tbl <- tibble::as_tibble(x) # x is lst, data.frame, tbl_df
+  tbl <- ungroup(as_tibble(x, validate = validate)) # x is lst, data.frame, tbl_df
   # extract or pass the index var
   index <- extract_index_var(tbl, index = index)
-  # validate key vars
+  # (1) validate and process key vars (from expr to a list of syms)
   key_vars <- validate_key(data = tbl, key)
+  # (2) if there exists a list of lists, flatten it as characters
   flat_keys <- flatten_key(key_vars)
-  is_index_in_keys <- intersect(quo_text2(index), flat_keys)
+  # (3) index cannot be part of the keys
+  idx_chr <- quo_text2(index)
+  is_index_in_keys <- intersect(idx_chr, flat_keys)
   if (is_false(is_empty(is_index_in_keys))) {
-    abort("The index variable cannot be one of the keys.")
+    msg <- sprintf(
+      "%s can't be both `key` and `index`.", surround(idx_chr, "`")
+    )
+    abort(msg)
   }
   # validate tbl_ts
   if (validate) {
@@ -386,11 +437,11 @@ tsibble_tbl <- function(x, key, index, regular = TRUE, validate = TRUE) {
     tbl_interval <- pull_interval(eval_idx, duplicated = TRUE)
   }
 
-  grped_key <- dplyr::grouped_df(tbl, flat_keys)
+  grped_key <- grouped_df(tbl, flat_keys)
   tibble::new_tibble(
-    tbl, 
+    tbl,
     "key" = structure(key_vars, class = "key"),
-    "key_indices" = group_size(grped_key),
+    "key_indices" = attr(grped_key, "indices"),
     "index" = index,
     "interval" = structure(tbl_interval, class = "interval"),
     "regular" = regular,
@@ -422,10 +473,10 @@ extract_index_var <- function(data, index) {
   if (quo_is_missing(index)) {
     val_idx <- idx_type %in% detect_type()
     if (sum(val_idx) != 1) {
-      abort("Please specify the 'index' argument.")
+      abort("Can't determine the `index`. Please specify the `index` arg.")
     }
     chr_index <- colnames(data)[val_idx]
-    inform(paste("The 'index' variable:", chr_index))
+    inform(sprintf("The `index` is %s.", surround(chr_index, "`")))
     idx_sym <- sym(chr_index)
     index <- as_quosure(idx_sym)
     return(index)
@@ -433,8 +484,8 @@ extract_index_var <- function(data, index) {
     idx_na <- idx_type[quo_text2(index)]
     if (is.na(idx_na)) {
       cls_idx <- purrr::map_chr(data, ~ class(.)[1])
-      abort(paste(
-        "Unsupported index type:",
+      abort(sprintf(
+        "Unsupported index type: %s",
         cls_idx[colnames(data) %in% names(idx_na)])
       )
     }
@@ -457,11 +508,12 @@ validate_nested <- function(data, key) {
     n_lgl <- purrr::map_lgl(n_dist, is_descending)
     if (is_false(all(n_lgl))) {
       which_bad <- key_nest[!n_lgl]
-      msg <- purrr::map(which_bad, ~ paste(., collapse = " | "))
-      msg <- paste(msg, collapse = ", ")
-      abort(paste0(
-        "Incorrect ordering of nested variables: ",
-        msg
+      wrong_nested <- purrr::map(which_bad, ~ paste(., collapse = " | "))
+      wrong_nested <- paste_comma(wrong_nested)
+      wrong_dim <- purrr::map_chr(n_dist, ~ paste(., collapse = " | "))
+      abort(sprintf(
+        "Incorrect nesting: %s %s. Please see `?tsibble`.",
+        wrong_nested, surround(wrong_dim, "(")
       ))
     }
   }
@@ -471,25 +523,42 @@ validate_nested <- function(data, key) {
 # check if a comb of key vars result in a unique data entry
 # if TRUE return the data, otherwise raise an error
 validate_tbl_ts <- function(data, key, index) {
-  tbl_dup <- dplyr::grouped_df(data, vars = flatten_key(key)) %>%
-    summarise(zzz = anyDuplicated(!! index))
+  idx <- quo_text2(index)
+  # NOTE: bug in anyDuplicated.data.frame()
+  # identifiers <- c(flatten_key(key), idx)
+  # below calls anyDuplicated.data.frame():
+  # time zone associated with the index will be dropped, 
+  # e.g. nycflights13::weather, thus result in duplicates. 
+  # dup <- anyDuplicated(data[, identifiers, drop = FALSE])
+  tbl_dup <- grouped_df(data, vars = flatten_key(key)) %>%
+    summarise(zzz = anyDuplicated.default(!! index))
   if (any_not_equal_to_c(tbl_dup$zzz, 0)) {
-    msg <- paste(
-      "Invalid tsibble with duplicated identifier:",
-      quo_text2(index)
-    )
+    msg <- sprintf("Invalid tsibble: identical data entries from %s", idx)
     if (!is_empty(key)) {
       class(key) <- "key"
-      msg <- paste(msg, "and", paste(format(key), collapse = ", "))
+      msg <- sprintf("%s and %s.", msg, paste_comma(format(key)))
     }
+    msg <- paste(msg, "Use `inform_duplicates()` to check the duplicated rows.")
     abort(msg)
   }
   data
 }
 
+#' Coerce to a tibble or data frame
+#'
+#' @param x A `tbl_ts`.
+#' @param ... Ignored.
+#'
+#' @rdname as-tibble
 #' @export
+#' @examples
+#' as_tibble(pedestrian)
+#'
+#' # a grouped tbl_ts -----
+#' grped_ped <- pedestrian %>% group_by(Sensor)
+#' as_tibble(grped_ped)
 as_tibble.tbl_ts <- function(x, ...) {
-  tibble::new_tibble(x)
+  drop_tsibble(x)
 }
 
 #' @export
@@ -497,22 +566,64 @@ as.tibble.tbl_ts <- as_tibble.tbl_ts
 
 #' @export
 as_tibble.grouped_ts <- function(x, ...) {
-  flat_grps <- flatten_key(groups(x))
-  dplyr::grouped_df(x, flat_grps)
+  dx <- drop_tsibble(x)
+  tibble::new_tibble(
+    dx,
+    "vars" = structure(flatten(groups(x)), class = "vars"),
+    subclass = "grouped_df"
+  )
 }
 
 #' @export
 as.tibble.grouped_ts <- as_tibble.grouped_ts
 
+#' @rdname as-tibble
 #' @export
-as.data.frame.tbl_ts <- function(x, row.names = NULL, optional = FALSE, ...) {
-  class(x) <- "data.frame"
-  x
+as.data.frame.tbl_ts <- function(x, ...) {
+  x <- drop_tsibble(x)
+  NextMethod()
 }
 
 use_id <- function(x) {
   tryCatch(
     is_quosures(x),
-    error = function(e) stop("Please use `tsibble::id()` to create the 'key'.")
+    error = function(e) {
+      e$call <- NULL
+      e$message <- "Have you forgotten `tsibble::id()` to create the `key`?"
+      stop(e)
+    }
   )
+}
+
+drop_tsibble <- function(x) {
+  attr(x, "key") <- attr(x, "key_indices") <- attr(x, "index") <- NULL
+  attr(x, "interval") <- attr(x, "regular") <- NULL
+  tibble::new_tibble(x)
+}
+
+#' Inform duplication of key and index variables
+#'
+#' Inform which row has duplicated key and index elements
+#'
+#' @param data A `tbl_ts` object.
+#' @param key Structural variable(s) that define unique time indices, used with
+#' the helper [id]. If a univariate time series (without an explicit key),
+#' simply call `id()`.
+#' @param index A bare (or unquoted) variable to specify the time index variable.
+#' @param fromLast `TRUE` does the duplication check from the last of identical
+#' elements.
+#'
+#' @return A logical vector of the same length as the row number of `data`
+#' @export
+inform_duplicates <- function(data, key = id(), index, fromLast = FALSE) {
+  use_id(key)
+  index <- extract_index_var(data, enquo(index))
+
+  grouped_df(data, vars = flatten_key(key)) %>%
+    mutate(zzz = duplicated.default(!! index)) %>% 
+    dplyr::pull(zzz)
+
+  # identifiers <- c(flatten_key(key), quo_text2(index))
+  # duplicated(data[, identifiers, drop = FALSE], fromLast = fromLast)
+  # not handling time zone correctly for duplicated.data.frame
 }