+++ R/dplyr-verbs.R
@@ -15,57 +15,29 @@
 #' @export
 arrange.tbl_ts <- function(.data, ...) {
   arr_data <- NextMethod()
-  as_tsibble(
-    arr_data, key = key(.data), index = !! index(.data),
-    validate = FALSE, regular = is_regular(.data)
-  )
+  update_tsibble(arr_data, .data)
 }
 
 #' @rdname row-verb
 #' @seealso [dplyr::arrange]
 #' @export
 arrange.grouped_ts <- function(.data, ..., .by_group = FALSE) {
-  grps <- groups(.data)
-  grped_data <- dplyr::grouped_df(.data, vars = flatten_key(grps))
-  arr_data <- arrange(grped_data, ..., .by_group = .by_group)
-  as_tsibble(
-    arr_data, key = key(.data), index = !! index(.data), groups = grps,
-    validate = FALSE, regular = is_regular(.data)
-  )
+  arr_data <- arrange(as_tibble(.data), ..., .by_group = .by_group)
+  update_tsibble(arr_data, .data)
 }
 
 #' @rdname row-verb
 #' @seealso [dplyr::filter]
 #' @export
 filter.tbl_ts <- function(.data, ...) {
-  grps <- groups(.data)
-  if (is_grouped_ts(.data)) {
-    grped_data <- dplyr::grouped_df(.data, vars = flatten_key(grps))
-    fil_data <- filter(grped_data, ...)
-  } else {
-    fil_data <- NextMethod()
-  }
-  as_tsibble(
-    fil_data, key = key(.data), index = !! index(.data), groups = grps,
-    validate = FALSE, regular = is_regular(.data)
-  )
+  by_row(filter, .data, ...)
 }
 
 #' @rdname row-verb
 #' @seealso [dplyr::slice]
 #' @export
 slice.tbl_ts <- function(.data, ...) {
-  grps <- groups(.data)
-  if (is_grouped_ts(.data)) {
-    grped_data <- dplyr::grouped_df(.data, vars = flatten_key(grps))
-    slc_data <- slice(grped_data, ...)
-  } else {
-    slc_data <- NextMethod()
-  }
-  as_tsibble(
-    slc_data, key = key(.data), index = !! index(.data), groups = grps,
-    validate = FALSE, regular = is_regular(.data)
-  )
+  by_row(slice, .data, ...)
 }
 
 #' Column-wise verbs
@@ -96,24 +68,24 @@ select.tbl_ts <- function(.data, ..., drop = FALSE) {
   val_vars <- validate_vars(j = lst_quos, x = colnames(.data))
   val_idx <- has_index_var(j = val_vars, x = .data)
   if (is_false(val_idx)) {
-    abort("The index variable cannot be dropped.")
+    abort(sprintf(
+      "The `index` (%s) must not be dropped. Do you need `drop = TRUE` to drop `tbl_ts`?", 
+      surround(quo_text2(index(.data)), "`")
+    ))
   }
   lhs <- names(val_vars)
   index(.data) <- update_index(index(.data), val_vars, lhs)
   val_key <- has_all_key(j = lst_quos, x = .data)
   if (is_true(val_key)) { # no changes in key vars
-    return(as_tsibble(
-      sel_data, key = key(.data), index = !! index(.data),
-      validate = FALSE, regular = is_regular(.data)
-    ))
-  } else {
-    key(.data) <- update_key(key(.data), val_vars)
-    key(.data) <- update_key2(key(.data), val_vars, lhs)
-    as_tsibble(
-      sel_data, key = key(.data), index = !! index(.data),
-      validate = TRUE, regular = is_regular(.data)
-    )
+    return(update_tsibble(sel_data, .data))
   }
+  key(.data) <- update_key(key(.data), val_vars)
+  key(.data) <- update_key2(key(.data), val_vars, lhs)
+  as_tsibble(
+    sel_data, key = key(.data), index = !! index(.data), 
+    groups = groups(.data), regular = is_regular(.data),
+    validate = TRUE,
+  )
 }
 
 #' @rdname col-verb
@@ -128,10 +100,7 @@ rename.tbl_ts <- function(.data, ...) {
     key(.data) <- update_key2(key(.data), val_vars, lhs)
   }
   ren_data <- NextMethod()
-  as_tsibble(
-    ren_data, key = key(.data), index = !! index(.data),
-    validate = FALSE, regular = is_regular(.data)
-  )
+  update_tsibble(ren_data, .data)
 }
 
 #' @rdname col-verb
@@ -144,11 +113,7 @@ mutate.tbl_ts <- function(.data, ..., drop = FALSE) {
   }
   lst_quos <- quos(..., .named = TRUE)
   vec_names <- union(names(lst_quos), colnames(.data))
-  grps <- groups(.data)
-  if (is_grouped_ts(.data)) {
-    grped_data <- dplyr::grouped_df(.data, vars = flatten_key(grps))
-    mut_data <- mutate(grped_data, ...)
-  }
+  mut_data <- mutate(as_tibble(.data), ...)
   # either key or index is present in ...
   # suggests that the operations are done on these variables
   # validate = TRUE to check if tsibble still holds
@@ -156,8 +121,9 @@ mutate.tbl_ts <- function(.data, ..., drop = FALSE) {
   val_key <- has_any_key(vec_names, .data)
   validate <- val_idx || val_key
   as_tsibble(
-    mut_data, key = key(.data), index = !! index(.data), groups = grps,
-    validate = validate, regular = is_regular(.data)
+    mut_data, key = key(.data), index = !! index(.data), 
+    groups = groups(.data), regular = is_regular(.data),
+    validate = validate 
   )
 }
 
@@ -169,20 +135,24 @@ summarise.tbl_ts <- function(.data, ..., drop = FALSE) {
     return(summarise(as_tibble(.data), ...))
   }
   lst_quos <- quos(..., .named = TRUE)
-  vec_vars <- as.character(purrr::map(lst_quos, ~ lang_args(.)[[1]]))
+  first_arg <- first_arg(lst_quos)
+  vec_vars <- as.character(first_arg)
+  idx <- index(.data)
   if (has_index_var(j = vec_vars, x = .data)) {
-    abort("The index variable cannot be summarised.")
+    abort(sprintf(
+      "The `index` (%s) must not be dropped. Do you need `drop = TRUE` to drop `tbl_ts`?", 
+      surround(quo_text2(idx), "`")
+    ))
   }
 
-  idx <- index(.data)
   grps <- groups(.data)
   chr_grps <- c(quo_text2(idx), flatten_key(grps))
   sum_data <- .data %>%
-    dplyr::grouped_df(vars = chr_grps) %>%
-    dplyr::summarise(!!! lst_quos)
+    grouped_df(vars = chr_grps) %>%
+    summarise(!!! lst_quos)
 
   as_tsibble(
-    sum_data, key = grps, index = !! idx, groups = grps,
+    sum_data, key = grps, index = !! idx, groups = drop_group(grps),
     validate = FALSE, regular = is_regular(.data)
   )
 }
@@ -204,6 +174,7 @@ summarize.tbl_ts <- summarise.tbl_ts
 #'
 #' @rdname group-by
 #' @seealso [dplyr::group_by]
+#' @importFrom dplyr grouped_df
 #' @export
 #' @examples
 #' data(tourism)
@@ -213,15 +184,19 @@ summarize.tbl_ts <- summarise.tbl_ts
 group_by.tbl_ts <- function(.data, ..., add = FALSE) {
   index <- index(.data)
   idx_var <- quo_text2(index)
-  grped_quo <- quos(...)
-  grped_chr <- prepare_groups(.data, grped_quo, add = add)
+  current_grps <- quos(...)
+  final_grps <- prepare_groups(.data, current_grps, add = add)
+  grped_chr <- flatten_key(final_grps)
   if (idx_var %in% grped_chr) {
-    abort(paste("The index variable", surround(idx_var), "cannot be grouped."))
+    abort(sprintf(
+      "The `index` (%s) must not be grouped. Do you need `as_tibble()` to coerce to `tbl_df`?", 
+      surround(idx_var, "`")
+    ))
   }
 
-  grped_ts <- grouped_ts(.data, grped_chr, grped_quo, add = add)
+  grped_ts <- grouped_df(.data, grped_chr)
   as_tsibble(
-    grped_ts, key = key(.data), index = !! index,
+    grped_ts, key = key(.data), index = !! index, groups = final_grps,
     validate = FALSE, regular = is_regular(.data)
   )
 }
@@ -230,10 +205,8 @@ group_by.tbl_ts <- function(.data, ..., add = FALSE) {
 #' @seealso [dplyr::ungroup]
 #' @export
 ungroup.grouped_ts <- function(x, ...) {
-  x <- rm_class(x, "grouped_ts")
-  groups(x) <- list()
   as_tsibble(
-    x, key = key(x), index = !! index(x),
+    x, key = key(x), index = !! index(x), groups = id(),
     validate = FALSE, regular = is_regular(x)
   )
 }
@@ -241,58 +214,43 @@ ungroup.grouped_ts <- function(x, ...) {
 #' @seealso [dplyr::ungroup]
 #' @export
 ungroup.tbl_ts <- function(x, ...) {
-  groups(x) <- list()
-  as_tsibble(
-    x, key = key(x), index = !! index(x),
-    validate = FALSE, regular = is_regular(x)
-  )
-}
-
-#' @importFrom dplyr group_indices
-#' @export
-group_indices.grouped_ts <- function(.data, ...) {
-  group_indices(as_tibble(.data))
+  x
 }
 
 # this function prepares group variables in a vector of characters for
 # dplyr::grouped_df()
-# the arg of group can take a nested str but be flattened in the end.
 prepare_groups <- function(data, group, add = FALSE) {
-  if (add) {
-    old_grps <- flatten_key(groups(data))
-    grps <- flatten_key(validate_key(data, group))
-    return(c(old_grps, grps))
-  } else {
-    flatten_key(validate_key(data, group))
+  grps <- validate_key(data, group)
+  if (is_false(add)) {
+    return(grps)
   }
+  c(groups(data), grps)
 }
 
-"groups<-" <- function(x, value) {
-  attr(x, "vars") <- value
-  x
+do.tbl_ts <- function(.data, ...) {
+  dplyr::do(as_tibble(.data), ...)
 }
 
-# work around with dplyr::grouped_df (not an elegant solution)
-# better to use dplyr internal cpp code when its API is stable
-# the arg `vars` passed to dplyr::grouped_df
-# the arg `group` takes new defined groups
-grouped_ts <- function(data, vars, group, add = FALSE) { # vars are characters
-  old_class <- class(data)
-  grped_df <- dplyr::grouped_df(data, vars)
-  class(grped_df) <- unique(c("grouped_ts", old_class))
-  val_grps <- validate_key(data, group)
-  data <- validate_nested(data = data, key = val_grps)
-  if (add) {
-    groups(grped_df) <- c(groups(data), val_grps)
-  } else {
-    groups(grped_df) <- val_grps
-  }
-  grped_df
+#' @export
+transmute.tbl_ts <- function(.data, ...) {
+  abort("'tbl_ts' has no support for transmute(). Please coerce to 'tbl_df' first and then transmute().")
 }
 
-# methods::setGeneric("groups<-")
+#' @export
+distinct.tbl_ts <- function(.data, ...) {
+  abort("'tbl_ts' has no support for distinct(). Please coerce to 'tbl_df' first and then distinct().")
+}
 
-rm_class <- function(x, value) {
-  class(x) <- class(x)[-match(value, class(x))]
-  x
+by_row <- function(FUN, .data, ...) {
+  FUN <- match.fun(FUN, descend = FALSE)
+  tbl <- FUN(as_tibble(.data), ...)
+  update_tsibble(tbl, .data)
+}
+
+# put new data with existing attributes
+update_tsibble <- function(new, old) {
+  as_tsibble(
+    new, key = key(old), index = !! index(old), groups = groups(old), 
+    regular = is_regular(old), validate = FALSE
+  )
 }