+++ R/parse-key.R
@@ -1,22 +1,40 @@
-reduce_key <- function(lst_keys) {
-  if (is_empty(lst_keys)) {
-    return(lst_keys)
+reduce_key <- function(x) { # x = a list of keys (symbols)
+  if (is_empty(x)) {
+    return(x)
   }
-  nest_lgl <- is_nest(lst_keys)
-  comb_keys <- lst_keys[!nest_lgl]
+  nest_lgl <- is_nest(x)
+  comb_keys <- x[!nest_lgl]
   if (any(nest_lgl)) {
-    nest_keys <- purrr::map(lst_keys[nest_lgl], ~ .[[1]])
+    nest_keys <- purrr::map(x[nest_lgl], ~ .[[1]])
     comb_keys <- c(nest_keys, comb_keys)
   }
   quos_auto_name(comb_keys)
 }
 
-# this returns a vector of groups/key characters
-flatten_key <- function(lst_keys) {
-  if (is.null(lst_keys)) {
-    lst_keys <- list()
+drop_group <- function(x) {
+  old_grps <- flatten_key(x)
+  len <- length(old_grps)
+  new_grps <- old_grps[-len] # drop one grouping level
+  lgl <- FALSE
+  if (!is_empty(x)) {
+    lgl <- rep(is_nest(x), purrr::map(x, length))
+    lgl <- lgl[-len]
   }
-  purrr::map_chr(flatten(lst_keys), quo_text2)
+  if (is_empty(new_grps)) {
+    return(id())
+  } else if (any(lgl)) {
+    return(c(list(syms(new_grps[lgl])), syms(new_grps[!lgl])))
+  } else {
+    syms(new_grps[!lgl])
+  }
+}
+
+# this returns a vector of groups/key characters
+flatten_key <- function(x) {
+  if (is.null(x)) {
+    x <- list()
+  }
+  purrr::map_chr(flatten(x), quo_text2)
 }
 
 # update by character names
@@ -83,12 +101,12 @@ is_nest <- function(lst_syms) {
   unname(purrr::map_lgl(lst_syms, is_list)) # expected to be a list not call
 }
 
-parse_key <- function(lst_keys) {
-  if (is_empty(lst_keys)) {
+parse_key <- function(x) {
+  if (is_empty(x)) {
     return(list())
   }
-  # purrr::map(lst_keys, ~ flatten_nest(.[[-1]]))
-  purrr::map(lst_keys, flatten_nest)
+  # purrr::map(x, ~ flatten_nest(.[[-1]]))
+  purrr::map(x, flatten_nest)
 }
 
 # interpret a nested calls A | B | C