+++ R/utils.R
@@ -28,6 +28,10 @@ paste_comma <- function(...) {
   paste(..., collapse = ", ")
 }
 
+first_arg <- function(x) {
+  purrr::compact(purrr::map(x, ~ dplyr::first(lang_args(.))))
+}
+
 # regular time interval is obtained from the minimal time distance.
 # duplicated time entries result in 0L.
 # if validate = FALSE in as_tsibble, skip to check duplicated entries
@@ -52,8 +56,10 @@ surround <- function(x, bracket = "(") {
     return(paste0("(", x, ")"))
   } else if (bracket == "[") {
     return(paste0("[", x, "]"))
-  } else {
+  } else if (bracket == "<") {
     paste0("<", x, ">")
+  } else {
+    paste0("`", x, "`")
   }
 }