+++ tests/testthat/test-dplyr.R
@@ -1,5 +1,25 @@
 library(lubridate)
-context("All dplyr verbs for tsibble")
+context("dplyr verbs for tsibble")
+
+test_that("group_by()", {
+  expect_error(group_by(tourism, State | Region))
+  expect_error(group_by(tourism, State | Region, Purpose))
+  expect_error(group_by(tourism, Quarter))
+  expect_error(group_by(pedestrian, Date_Time, Sensor))
+  grped_df <- pedestrian %>% 
+    group_by(Date) %>% 
+    group_by(Sensor, add = TRUE)
+  expect_length(group_vars(grped_df), 2)
+  grped_df <- pedestrian %>% 
+    group_by(Sensor)
+  expect_equal(n_groups(grped_df), 4)
+  expect_length(group_size(grped_df), 4)
+
+  grped_t <- tourism %>% 
+    group_by(Purpose) %>% 
+    group_by(Region | State, add = TRUE)
+  expect_length(group_vars(grped_t), 2)
+})
 
 test_that("arrange()", {
  tsbl1 <- arrange(tourism, Quarter)
@@ -23,8 +43,8 @@ test_that("filter() and slice()", {
   tsbl2 <- tourism %>%
     group_by(!!! key(.)) %>%
     filter(Quarter < yearquarter(ymd("1999-01-01")))
-  # expect_identical(key(tsbl2), key(tourism))
-  # expect_identical(group_vars(tsbl2), key_vars(tourism))
+  expect_identical(key(tsbl2), key(tourism))
+  expect_identical(group_vars(tsbl2), key_vars(tourism))
   expect_identical(index(tsbl2), index(tourism))
   tsbl3 <- slice(tourism, 1:3)
   expect_identical(dim(tsbl3), c(3L, ncol(tourism)))
@@ -38,27 +58,27 @@ test_that("select() and rename()", {
   expect_error(select(tourism, Quarter))
   expect_error(select(tourism, Region))
   expect_is(select(tourism, Region, drop = TRUE), "tbl_df")
-  # expect_is(select(tourism, Quarter:Purpose), "tbl_ts")
-  # expect_equal(
-  #   quo_name(index(select(tourism, Index = Quarter, Region:Purpose))),
-  #   "Index"
-  # )
-  # expect_equal(
-  #   quo_name(index(select(tourism, Index = Quarter, Region:Purpose))),
-  #   "Index"
-  # )
-  # expect_equal(
-  #   key_vars(select(tourism, Bottom = Region, Quarter, State:Purpose))[[1]],
-  #   "Bottom | State"
-  # )
-  # expect_equal(
-  #   quo_name(index(rename(tourism, Index = Quarter))),
-  #   "Index"
-  # )
-  # expect_equal(
-  #   key_vars(rename(tourism, Bottom = Region))[[1]],
-  #   "Bottom | State"
-  # )
+  expect_is(select(tourism, Quarter:Purpose), "tbl_ts")
+  expect_equal(
+    quo_name(index(select(tourism, Index = Quarter, Region:Purpose))),
+    "Index"
+  )
+  expect_equal(
+    quo_name(index(select(tourism, Index = Quarter, Region:Purpose))),
+    "Index"
+  )
+  expect_equal(
+    key_vars(select(tourism, Bottom = Region, Quarter, State:Purpose))[[1]],
+    "Bottom | State"
+  )
+  expect_equal(
+    quo_name(index(rename(tourism, Index = Quarter))),
+    "Index"
+  )
+  expect_equal(
+    key_vars(rename(tourism, Bottom = Region))[[1]],
+    "Bottom | State"
+  )
 })
 
 test_that("mutate()", {
@@ -85,8 +105,22 @@ test_that("summarise()", {
     group_by(!!! key(.)) %>%
     summarise(Trips = sum(Trips))
   expect_equal(tourism, tsbl2)
-  # expect_identical(key(tourism), key(tsbl2))
+  expect_identical(key(tourism), key(tsbl2))
   expect_identical(index(tourism), index(tsbl2))
   expect_identical(is_regular(tourism), is_regular(tsbl2))
-  # expect_identical(group_vars(tsbl2), key_vars(tourism))
+  tsbl3 <- tourism %>%
+    group_by(!!! key(.)) %>%
+    summarise(Obs = n())
+  expect_identical(nrow(tsbl3), nrow(tourism))
+
+  expect_error(pedestrian %>% summarise(month = yearmonth(Date_Time)))
+  tbl_ped <- pedestrian %>% 
+    group_by(Date) %>% 
+    summarise(DailyCount = mean(Count), drop = TRUE)
+  expect_is(tbl_ped, "tbl_df")
+})
+
+test_that("transmute() and distinct()", {
+  expect_error(tourism %>% transmute(Region = paste(Region, State)))
+  expect_error(tourism %>% distinct(Region, State, Purpose))
 })